+++ inst/examples/update_app/app.R
@@ -1,23 +1,68 @@
 library(shiny)
 library(shinyMobile)
 
+colors <- c(
+  lightblue = "#5ac8fa",
+  pink = "#ff2d55",
+  teal = "#009688",
+  yellow = "#ffcc00"
+)
+
 app <- shinyApp(
   ui = f7Page(
-    title = "Simple Dialog",
+    title = "Update App",
+    options = (
+      list(
+        color = "#5ac8fa"
+      )
+    ),
     f7SingleLayout(
-      navbar = f7Navbar(title = "f7Dialog"),
+      navbar = f7Navbar(title = "Update App"),
+      f7BlockTitle("Update f7Dialog configuration"),
       f7Segment(
         f7Button(
           inputId = "goButton",
-          "Go!",
-          fill = FALSE,
-          outline = TRUE
+          "Show f7Dialog"
         ),
         f7Button(
           inputId = "update",
-          "Update config",
-          fill = FALSE,
-          outline = TRUE
+          "Update config"
+        )
+      ),
+      f7BlockTitle("Update theme"),
+      f7Segment(
+        f7Button(
+          inputId = "theme_ios",
+          "iOS theme"
+        ),
+        f7Button(
+          inputId = "theme_md",
+          "MD theme"
+        )
+      ),
+      f7BlockTitle("Set dark mode"),
+      f7Segment(
+        f7Button(
+          inputId = "enable_darkmode",
+          "Enable darkmode"
+        ),
+        f7Button(
+          inputId = "disable_darkmode",
+          "Disable darkmode"
+        )
+      ),
+      f7BlockTitle("Change color theme"),
+      f7Segment(
+        tagList(
+          lapply(names(colors),
+                 function(color) {
+                   f7Button(
+                     inputId = paste0("color_", color),
+                     label = color,
+                     color = color,
+                   )
+                 }
+          )
         )
       )
     )
@@ -33,7 +78,6 @@ app <- shinyApp(
     observeEvent(input$update, {
       updateF7App(
         options = list(
-          theme = "ios",
           dialog = list(
             buttonOk = "Yeaaaah!",
             buttonCancel = "Ouuups!"
@@ -48,6 +92,49 @@ app <- shinyApp(
         text = "Look at me, I have a new buttons!"
       )
     })
+
+    observeEvent(input$theme_ios, {
+      updateF7App(
+        options = list(
+          theme = "ios"
+        )
+      )
+    })
+
+    observeEvent(input$theme_md, {
+      updateF7App(
+        options = list(
+          theme = "md"
+        )
+      )
+    })
+
+    observeEvent(input$enable_darkmode, {
+      updateF7App(
+        options = list(
+          dark = TRUE
+        )
+      )
+    })
+
+    observeEvent(input$disable_darkmode, {
+      updateF7App(
+        options = list(
+          dark = FALSE
+        )
+      )
+    })
+
+    lapply(names(colors), function(color) {
+      observeEvent(input[[paste0("color_", color)]], {
+        updateF7App(
+          options = list(
+            color = colors[color]
+          )
+        )
+      })
+    })
+
   }
 )