+++ NEWS.md
@@ -87,6 +87,7 @@ the new router layout. Items must be wrapped in a `shiny::tagList()`.
 - `f7Picker`, `f7DatePicker` and `f7ColorPicker` now have an additional `style` parameter to set the input style and inherit from `f7List` options such as outline, inset, strong and dividers.
 - `f7Toggle()` and `f7Stepper()` automatically detect when they are called inside a `f7List` and adjust accordingly to the list style.
 - `f7TabLink()`: `href` can now be different of `#` if passed in `...` (Default is still `#` when no custom `href` is detected). This allows to use tab links as routable links but with a better rendering in the toolbar/tabbar than classic `f7Link()`.
+- `updateF7App` can now also handle changes in app theme (ios or md), dark mode, and color.
 - Fix various issues in documentation.
 
 # shinyMobile 1.0.1